Splitting server into multiple files.
